Fix parallel rendering
Creating one worker per chunk is very inefficient. It's much better to
create N workers (typically the same as the number of logical cores)
so that these cores don't spend time thrashing.

Add in timing information for each worker and also increase rendering
delay to 1/4 of a second (1/10 of a second was too short).